fix: Complete database schema alignment and cleanup
Major fixes and cleanup:
- Fixed critical schema mismatches between models and migration
  - gene_normalization_staging: source -> source_name
  - gene_normalization_log: added missing source_name column
  - All foreign keys now use BigInteger consistently
- Removed all obsolete migration folders and scripts
  - Deleted 28 old migration files from archived/backup folders
  - Removed 6 migration refactor scripts (no longer needed)
  - Removed setup_database.py (replaced by Alembic)
- Kept only essential utility scripts with documentation
- Created comprehensive README for scripts folder

Current state:
- Single clean migration: 001_modern_complete_schema.py
- All models aligned with database schema
- All endpoints working (staging, progress, logs)
- 1,697 genes loaded successfully
- Code fully linted and clean

+++ backend/app/models/cache.py
@@ -2,11 +2,10 @@
 Cache model for database-backed caching
 """
 
-import uuid
 from datetime import datetime
 
-from sqlalchemy import DateTime, Index, Integer, Text, UniqueConstraint
-from sqlalchemy.dialects.postgresql import JSONB, UUID
+from sqlalchemy import BigInteger, DateTime, Index, Integer, Text, UniqueConstraint
+from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import Mapped, mapped_column
 from sqlalchemy.sql import func
 
@@ -20,7 +19,7 @@ class CacheEntry(Base):
 
     # Fix: Database has INTEGER id, not UUID
     id: Mapped[int] = mapped_column(
-        Integer, primary_key=True, index=True
+        BigInteger, primary_key=True, index=True
     )
     # Fix: Database has VARCHAR(255), but we can safely use Text in model
     cache_key: Mapped[str] = mapped_column(Text, nullable=False)
